Create tapView that overlays the raindrops.
When raindrops are added to the main view, taps near or on the created raindrop will not produce new raindrops
because the view being tapped on is the raindrop instead of the view which has the tap gesture recognizer. This issue
is solved by creating a tap view that overlays everything making the user only interact with the tap view.